Move httputil/responseutil to adapters/util Move adapters/openrtb_util to exchange/util
