+++ adapters/flipp/flipp.go
@@ -11,9 +11,9 @@ import (
 	"github.com/gofrs/uuid"
 	"github.com/prebid/openrtb/v19/openrtb2"
 	"github.com/prebid/prebid-server/adapters"
+	"github.com/prebid/prebid-server/adapters/util"
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/openrtb_ext"
-	"github.com/prebid/prebid-server/util/httputil"
 )
 
 const (
@@ -174,11 +174,11 @@ func buildPrebidRequest(flippExtParams openrtb_ext.ImpExtFlipp, request *openrtb
 }
 
 func (a *adapter) MakeBids(request *openrtb2.BidRequest, requestData *adapters.RequestData, responseData *adapters.ResponseData) (*adapters.BidderResponse, []error) {
-	if httputil.IsResponseStatusCodeNoContent(responseData) {
+	if util.IsResponseStatusCodeNoContent(responseData) {
 		return nil, nil
 	}
 
-	if err := httputil.CheckResponseStatusCodeForErrors(responseData); err != nil {
+	if err := util.CheckResponseStatusCodeForErrors(responseData); err != nil {
 		return nil, []error{err}
 	}
