+++ adapters/frvradn/frvradn.go
@@ -4,13 +4,14 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"strings"
+
 	"github.com/prebid/openrtb/v19/openrtb2"
 	"github.com/prebid/prebid-server/adapters"
+	"github.com/prebid/prebid-server/adapters/util"
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/errortypes"
 	"github.com/prebid/prebid-server/openrtb_ext"
-	"github.com/prebid/prebid-server/util/httputil"
-	"strings"
 )
 
 type adapter struct {
@@ -76,11 +77,11 @@ func (a *adapter) MakeRequests(request *openrtb2.BidRequest, requestInfo *adapte
 }
 
 func (a *adapter) MakeBids(request *openrtb2.BidRequest, requestData *adapters.RequestData, responseData *adapters.ResponseData) (*adapters.BidderResponse, []error) {
-	if httputil.IsResponseStatusCodeNoContent(responseData) {
+	if util.IsResponseStatusCodeNoContent(responseData) {
 		return nil, nil
 	}
 
-	if err := httputil.CheckResponseStatusCodeForErrors(responseData); err != nil {
+	if err := util.CheckResponseStatusCodeForErrors(responseData); err != nil {
 		return nil, []error{err}
 	}
