+++ adapters/util/response_test.go
@@ -1,4 +1,4 @@
-package httputil
+package util
 
 import (
 	"testing"
